\name{DPBBM-package}
\alias{DPBBM-package}
\alias{DPBBM}
\docType{package}
\title{
DPBBM
}
\description{
This package is developed for the beta-binomial mixture model based clustering
}
\details{
\tabular{ll}{
Package: \tab DPBBM\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2016-06-26\cr
License: \tab GPL-2 \cr
}
}

\references{

Coming soon!

}

\examples{
# Please check the main function of the package
?dpbbm_mc_iterations 
}